DEV | Headless | Date | TIme
In this update of Cobra Client I have shut down a few features of the Client due to them not working correctly, they will be fixed soon, but for now they are not working, and for the developers of the Client we are working on the Developer Menu of the client. Thank you for choosing to use Cobra Client.